Refactoring de métodos.
Se separó la implementación de métodos de carga e inserción de usuarios.
