+++ src/main/java/ar/com/buildingways/agenciapp/dao/UserRepositoryCustom.java
@@ -1,0 +1,11 @@
+package ar.com.buildingways.agenciapp.dao;
+
+import java.util.Collection;
+
+public interface UserRepositoryCustom {
+	
+	Collection<Object[]> loadUsersFromDB();
+	
+	void insertUsers(Collection<Object[]> usersFromDB);
+
+}
