+++ src/main/java/ar/com/buildingways/agenciapp/repository/UserRepository.java
@@ -1,12 +1,11 @@
 package ar.com.buildingways.agenciapp.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
+import ar.com.buildingways.agenciapp.dao.UserRepositoryCustom;
 import ar.com.buildingways.agenciapp.model.User;
 
-@Repository("userRepository")
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Integer>, UserRepositoryCustom {
 
 	User findByUsername(int username);
